fix: support specs with platform-specific transitives
This supports the case where the pod is not a direct dependency of any targets of the integration, which can happen for platform-specific transitives, when their platform is not used in any target. (e.g. PromiseKit/UIKit is iOS-specific and is a transitive of PromiseKit, but won't be included for a macOS-only integration.)